Unbreak the Travis tests
sov.pdf consistently builds as 5c52eea... on my system, but
consistently builds as 49bfca2... inside docker.

Update the SHA256SUMS file by hand for now; however, to fix this
properly, we need to be able to run update-test-expectation.sh
inside Docker (or update the script to use docker run and docker cp).